+++ library/core/src/main/java/com/google/android/exoplayer2/DefaultRenderersFactory.java
@@ -20,14 +20,11 @@ import android.media.MediaCodec;
 import android.os.Handler;
 import android.os.Looper;
 import androidx.annotation.IntDef;
-import androidx.annotation.Nullable;
 import com.google.android.exoplayer2.audio.AudioCapabilities;
 import com.google.android.exoplayer2.audio.AudioProcessor;
 import com.google.android.exoplayer2.audio.AudioRendererEventListener;
 import com.google.android.exoplayer2.audio.DefaultAudioSink;
 import com.google.android.exoplayer2.audio.MediaCodecAudioRenderer;
-import com.google.android.exoplayer2.drm.DrmSessionManager;
-import com.google.android.exoplayer2.drm.FrameworkMediaCrypto;
 import com.google.android.exoplayer2.mediacodec.MediaCodecRenderer;
 import com.google.android.exoplayer2.mediacodec.MediaCodecSelector;
 import com.google.android.exoplayer2.metadata.MetadataOutput;
@@ -88,10 +85,8 @@ public class DefaultRenderersFactory implements RenderersFactory {
   protected static final int MAX_DROPPED_VIDEO_FRAME_COUNT_TO_NOTIFY = 50;
 
   private final Context context;
-  @Nullable private DrmSessionManager<FrameworkMediaCrypto> drmSessionManager;
   @ExtensionRendererMode private int extensionRendererMode;
   private long allowedVideoJoiningTimeMs;
-  private boolean playClearSamplesWithoutKeys;
   private boolean enableDecoderFallback;
   private MediaCodecSelector mediaCodecSelector;
   @MediaCodecRenderer.MediaCodecOperationMode private int mediaCodecOperationMode;
@@ -106,17 +101,6 @@ public class DefaultRenderersFactory implements RenderersFactory {
   }
 
   /**
-   * @deprecated Use {@link #DefaultRenderersFactory(Context)} and pass {@link DrmSessionManager}
-   *     directly to {@link SimpleExoPlayer.Builder}.
-   */
-  @Deprecated
-  @SuppressWarnings("deprecation")
-  public DefaultRenderersFactory(
-      Context context, @Nullable DrmSessionManager<FrameworkMediaCrypto> drmSessionManager) {
-    this(context, drmSessionManager, EXTENSION_RENDERER_MODE_OFF);
-  }
-
-  /**
    * @deprecated Use {@link #DefaultRenderersFactory(Context)} and {@link
    *     #setExtensionRendererMode(int)}.
    */
@@ -128,47 +112,17 @@ public class DefaultRenderersFactory implements RenderersFactory {
   }
 
   /**
-   * @deprecated Use {@link #DefaultRenderersFactory(Context)} and {@link
-   *     #setExtensionRendererMode(int)}, and pass {@link DrmSessionManager} directly to {@link
-   *     SimpleExoPlayer.Builder}.
-   */
-  @Deprecated
-  @SuppressWarnings("deprecation")
-  public DefaultRenderersFactory(
-      Context context,
-      @Nullable DrmSessionManager<FrameworkMediaCrypto> drmSessionManager,
-      @ExtensionRendererMode int extensionRendererMode) {
-    this(context, drmSessionManager, extensionRendererMode, DEFAULT_ALLOWED_VIDEO_JOINING_TIME_MS);
-  }
-
-  /**
    * @deprecated Use {@link #DefaultRenderersFactory(Context)}, {@link
    *     #setExtensionRendererMode(int)} and {@link #setAllowedVideoJoiningTimeMs(long)}.
    */
   @Deprecated
-  @SuppressWarnings("deprecation")
   public DefaultRenderersFactory(
       Context context,
-      @ExtensionRendererMode int extensionRendererMode,
-      long allowedVideoJoiningTimeMs) {
-    this(context, null, extensionRendererMode, allowedVideoJoiningTimeMs);
-  }
-
-  /**
-   * @deprecated Use {@link #DefaultRenderersFactory(Context)}, {@link
-   *     #setExtensionRendererMode(int)} and {@link #setAllowedVideoJoiningTimeMs(long)}, and pass
-   *     {@link DrmSessionManager} directly to {@link SimpleExoPlayer.Builder}.
-   */
-  @Deprecated
-  public DefaultRenderersFactory(
-      Context context,
-      @Nullable DrmSessionManager<FrameworkMediaCrypto> drmSessionManager,
       @ExtensionRendererMode int extensionRendererMode,
       long allowedVideoJoiningTimeMs) {
     this.context = context;
     this.extensionRendererMode = extensionRendererMode;
     this.allowedVideoJoiningTimeMs = allowedVideoJoiningTimeMs;
-    this.drmSessionManager = drmSessionManager;
     mediaCodecSelector = MediaCodecSelector.DEFAULT;
   }
 
@@ -200,25 +154,6 @@ public class DefaultRenderersFactory implements RenderersFactory {
   public DefaultRenderersFactory experimental_setMediaCodecOperationMode(
       @MediaCodecRenderer.MediaCodecOperationMode int mode) {
     mediaCodecOperationMode = mode;
-    return this;
-  }
-
-  /**
-   * Sets whether renderers are permitted to play clear regions of encrypted media prior to having
-   * obtained the keys necessary to decrypt encrypted regions of the media. For encrypted media that
-   * starts with a short clear region, this allows playback to begin in parallel with key
-   * acquisition, which can reduce startup latency.
-   *
-   * <p>The default value is {@code false}.
-   *
-   * @param playClearSamplesWithoutKeys Whether renderers are permitted to play clear regions of
-   *     encrypted media prior to having obtained the keys necessary to decrypt encrypted regions of
-   *     the media.
-   * @return This factory, for convenience.
-   */
-  public DefaultRenderersFactory setPlayClearSamplesWithoutKeys(
-      boolean playClearSamplesWithoutKeys) {
-    this.playClearSamplesWithoutKeys = playClearSamplesWithoutKeys;
     return this;
   }
 
@@ -269,18 +204,12 @@ public class DefaultRenderersFactory implements RenderersFactory {
       VideoRendererEventListener videoRendererEventListener,
       AudioRendererEventListener audioRendererEventListener,
       TextOutput textRendererOutput,
-      MetadataOutput metadataRendererOutput,
-      @Nullable DrmSessionManager<FrameworkMediaCrypto> drmSessionManager) {
-    if (drmSessionManager == null) {
-      drmSessionManager = this.drmSessionManager;
-    }
+      MetadataOutput metadataRendererOutput) {
     ArrayList<Renderer> renderersList = new ArrayList<>();
     buildVideoRenderers(
         context,
         extensionRendererMode,
         mediaCodecSelector,
-        drmSessionManager,
-        playClearSamplesWithoutKeys,
         enableDecoderFallback,
         eventHandler,
         videoRendererEventListener,
@@ -290,8 +219,6 @@ public class DefaultRenderersFactory implements RenderersFactory {
         context,
         extensionRendererMode,
         mediaCodecSelector,
-        drmSessionManager,
-        playClearSamplesWithoutKeys,
         enableDecoderFallback,
         buildAudioProcessors(),
         eventHandler,
@@ -312,11 +239,6 @@ public class DefaultRenderersFactory implements RenderersFactory {
    * @param context The {@link Context} associated with the player.
    * @param extensionRendererMode The extension renderer mode.
    * @param mediaCodecSelector A decoder selector.
-   * @param drmSessionManager An optional {@link DrmSessionManager}. May be null if the player will
-   *     not be used for DRM protected playbacks.
-   * @param playClearSamplesWithoutKeys Whether renderers are permitted to play clear regions of
-   *     encrypted media prior to having obtained the keys necessary to decrypt encrypted regions of
-   *     the media.
    * @param enableDecoderFallback Whether to enable fallback to lower-priority decoders if decoder
    *     initialization fails. This may result in using a decoder that is slower/less efficient than
    *     the primary decoder.
@@ -330,8 +252,6 @@ public class DefaultRenderersFactory implements RenderersFactory {
       Context context,
       @ExtensionRendererMode int extensionRendererMode,
       MediaCodecSelector mediaCodecSelector,
-      @Nullable DrmSessionManager<FrameworkMediaCrypto> drmSessionManager,
-      boolean playClearSamplesWithoutKeys,
       boolean enableDecoderFallback,
       Handler eventHandler,
       VideoRendererEventListener eventListener,
@@ -342,8 +262,6 @@ public class DefaultRenderersFactory implements RenderersFactory {
             context,
             mediaCodecSelector,
             allowedVideoJoiningTimeMs,
-            drmSessionManager,
-            playClearSamplesWithoutKeys,
             enableDecoderFallback,
             eventHandler,
             eventListener,
@@ -420,11 +338,6 @@ public class DefaultRenderersFactory implements RenderersFactory {
    * @param context The {@link Context} associated with the player.
    * @param extensionRendererMode The extension renderer mode.
    * @param mediaCodecSelector A decoder selector.
-   * @param drmSessionManager An optional {@link DrmSessionManager}. May be null if the player will
-   *     not be used for DRM protected playbacks.
-   * @param playClearSamplesWithoutKeys Whether renderers are permitted to play clear regions of
-   *     encrypted media prior to having obtained the keys necessary to decrypt encrypted regions of
-   *     the media.
    * @param enableDecoderFallback Whether to enable fallback to lower-priority decoders if decoder
    *     initialization fails. This may result in using a decoder that is slower/less efficient than
    *     the primary decoder.
@@ -438,8 +351,6 @@ public class DefaultRenderersFactory implements RenderersFactory {
       Context context,
       @ExtensionRendererMode int extensionRendererMode,
       MediaCodecSelector mediaCodecSelector,
-      @Nullable DrmSessionManager<FrameworkMediaCrypto> drmSessionManager,
-      boolean playClearSamplesWithoutKeys,
       boolean enableDecoderFallback,
       AudioProcessor[] audioProcessors,
       Handler eventHandler,
@@ -449,8 +360,6 @@ public class DefaultRenderersFactory implements RenderersFactory {
         new MediaCodecAudioRenderer(
             context,
             mediaCodecSelector,
-            drmSessionManager,
-            playClearSamplesWithoutKeys,
             enableDecoderFallback,
             eventHandler,
             eventListener,